refactor(js): implement ES6 syntax and module system
