+++ index.html
@@ -20,9 +20,8 @@
     <main>
         <h2> INSTRUCTIONS</h2>
         <h3> My game is the best game. It is the greatest game ever known to man. Please play my game. You won't be dissapointed!</h3>
-        <p> Concentration, also known as Match Match, Memory, Pelmanism, Shinkei-suijaku, Pexeso or simply Pairs, is a card game
-            in which all of the cards are laid face down on a surface and two cards are flipped face up over each turn. The
-            object of the game is to turn over pairs of matching cards.</p>
+        <p> Concentration, also known as Match Match, Memory, Pelmanism, Shinkei-suijaku, Pexeso or simply Pairs, is a card game in which all of the cards are laid face down on a surface and two cards are flipped face up over each turn. The object of the
+            game is to turn over pairs of matching cards.</p>
         <div id="game-board" class="board clearfix"></div>
     </main>
     <footer class="clearfix">
@@ -31,7 +30,7 @@
             <span class="name">GA</span>
         </p>
     </footer>
-    <script src="js/main.js"></script>
+    <script type="module" src="js/main.js"></script>
 </body>
 
 </html>
